fix(api): handle undecryptable connection fields after key change
